# docs/openapi.yaml
openapi: 3.1.0
info:
  title: shipyard-cp API
  version: 0.1.0
  summary: AI worker orchestration control plane API draft
  description: |
    shipyard-cp の Task / WorkerJob / WorkerResult / StateTransitionEvent を扱う最小 API 契約。
    LiteLLM を推論ゲートウェイ、Codex / Claude Code / Google Antigravity をワーカーとして扱う。
    internal task state / typed_ref / context bundle は agent-taskstate、tracker 連携は tracker-bridge-materials、
    docs / contract resolve と read ack は memx-resolver へ委譲する。
servers:
  - url: https://api.shipyard-cp.local
    description: Production placeholder
  - url: http://localhost:3000
    description: Local development
jsonSchemaDialect: https://json-schema.org/draft/2020-12/schema
paths:
  /v1/tasks:
    post:
      operationId: createTask
      summary: Task を作成する
      description: agent-taskstate と整合する canonical typed_ref を持つ Task を作成する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/tasks/{task_id}:
    get:
      operationId: getTask
      summary: Task を取得する
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/tasks/{task_id}/dispatch:
    post:
      operationId: dispatchTaskJob
      summary: 次工程の WorkerJob を生成する
      description: Control Plane が Task state と resolver / tracker context をもとに WorkerJob を生成する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerJob'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/results:
    post:
      operationId: submitWorkerResult
      summary: WorkerResult を反映する
      description: ワーカーの結果を正規化し、Task / audit event / external refs へ反映する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerResult'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultApplyResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/transitions:
    post:
      operationId: recordTaskTransition
      summary: 手動またはポリシー起因の状態遷移を記録する
      description: 許可遷移は state-machine.md の許可遷移一覧、および StateTransitionEvent schema の oneOf 制約に一致しなければならない。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateTransitionEvent'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateTransitionEvent'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/docs/resolve:
    post:
      operationId: resolveTaskDocs
      summary: memx-resolver で読むべき文書を解決する
      description: feature / topic / task seed から doc refs と chunk refs を解決する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveDocsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveDocsResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/docs/ack:
    post:
      operationId: ackTaskDocs
      summary: memx-resolver へ読了記録を残す
      description: task context が参照した文書について read ack を記録する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckDocsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckDocsResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/tracker/link:
    post:
      operationId: linkTaskTrackerEntity
      summary: tracker entity を Task に紐付ける
      description: tracker-bridge-materials を通して tracker issue / project item / sync event を内部 task に接続する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackerLinkRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerLinkResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/integrate:
    post:
      operationId: integrateTask
      summary: Integrate を開始する
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrateRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrateResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/publish:
    post:
      operationId: publishTask
      summary: Publish を開始する
      description: published を終端状態とし、No-op / Dry-run / Apply はいずれも最終的に published へ到達する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/cancel:
    post:
      operationId: cancelTask
      summary: Task を中止する
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/events:
    get:
      operationId: listTaskEvents
      summary: Task の状態遷移イベント一覧を取得する
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/StateTransitionEvent'
  /v1/tasks/{task_id}/acceptance/complete:
    post:
      operationId: completeAcceptance
      summary: 手動受入を完了する
      description: accepting 状態の Task について、チェックリスト検証と verdict 確認を経て accepted へ遷移する。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteAcceptanceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteAcceptanceResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/tasks/{task_id}/audit-events:
    get:
      operationId: listTaskAuditEvents
      summary: Task の監査イベント一覧を取得する
      description: run.main_updated, run.publishRequested, run.publishCompleted, task.verdictSubmitted, run.permissionEscalated などの監査イベントを返す。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
  /v1/jobs/{job_id}:
    get:
      operationId: getJob
      summary: WorkerJob または最新状態を取得する
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/jobs/{job_id}/heartbeat:
    post:
      operationId: heartbeatJob
      summary: 実行中 worker job の heartbeat を更新する
      description: worker-dispatched stages の active job に対して heartbeat を反映する。integrate / publish の進行監視は Control Plane 内部イベントで扱ってよい。
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
components:
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: State conflict or validation gate failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Gone:
      description: Resource is no longer active
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Task:
      $ref: ./schemas/task.schema.json
    WorkerJob:
      $ref: ./schemas/worker-job.schema.json
    WorkerResult:
      $ref: ./schemas/worker-result.schema.json
    StateTransitionEvent:
      $ref: ./schemas/state-transition-event.schema.json
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [title, objective, typed_ref, repo_ref]
      properties:
        title:
          type: string
          minLength: 1
        objective:
          type: string
          minLength: 1
        typed_ref:
          type: string
          pattern: ^[a-z0-9_-]+:[a-z0-9_-]+:[a-z0-9_-]+:.+$
        description:
          type: string
        repo_ref:
          type: object
          additionalProperties: false
          required: [provider, owner, name, default_branch]
          properties:
            provider:
              type: string
              enum: [github]
            owner:
              type: string
              minLength: 1
            name:
              type: string
              minLength: 1
            default_branch:
              type: string
              minLength: 1
            base_sha:
              type: string
              pattern: ^[a-fA-F0-9]{7,64}$
        risk_level:
          type: string
          enum: [low, medium, high]
        labels:
          type: array
          items:
            type: string
          uniqueItems: true
        external_refs:
          type: array
          items:
            $ref: '#/components/schemas/ExternalRef'
        publish_plan:
          type: object
          additionalProperties: false
          properties:
            mode:
              type: string
              enum: [no_op, dry_run, apply]
            idempotency_key:
              type: string
            targets:
              type: array
              items:
                type: string
                enum: [deployment, release, package_publish, external_api]
              uniqueItems: true
            approval_required:
              type: boolean
    ExternalRef:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties:
        kind:
          type: string
          enum: [github_issue, github_project_item, release, deployment, tag, tracker_issue, sync_event, entity_link]
        value:
          type: string
          minLength: 1
        connection_ref:
          type: string
    DispatchRequest:
      type: object
      additionalProperties: false
      required: [target_stage]
      properties:
        target_stage:
          type: string
          enum: [plan, dev, acceptance]
        worker_selection:
          type: string
          enum: [codex, claude_code, google_antigravity]
        override_risk_level:
          type: string
          enum: [low, medium, high]
        expected_version:
          type: integer
          minimum: 0
    ResolveDocsRequest:
      type: object
      additionalProperties: false
      properties:
        feature:
          type: string
        topic:
          type: string
        task_seed:
          type: string
    ResolveDocsResponse:
      type: object
      additionalProperties: false
      required: [typed_ref, doc_refs, chunk_refs, contract_refs, stale_status]
      properties:
        typed_ref:
          type: string
          pattern: ^[a-z0-9_-]+:[a-z0-9_-]+:[a-z0-9_-]+:.+$
        doc_refs:
          type: array
          items:
            type: string
          uniqueItems: true
        chunk_refs:
          type: array
          items:
            type: string
          uniqueItems: true
        contract_refs:
          type: array
          items:
            type: string
          uniqueItems: true
        stale_status:
          type: string
          enum: [fresh, stale, unknown]
    AckDocsRequest:
      type: object
      additionalProperties: false
      required: [doc_id, version, task_id]
      properties:
        doc_id:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        task_id:
          type: string
          minLength: 1
    AckDocsResponse:
      type: object
      additionalProperties: false
      required: [ack_ref]
      properties:
        ack_ref:
          type: string
          minLength: 1
    TrackerLinkRequest:
      type: object
      additionalProperties: false
      required: [typed_ref, connection_ref, entity_ref]
      properties:
        typed_ref:
          type: string
          pattern: ^[a-z0-9_-]+:[a-z0-9_-]+:[a-z0-9_-]+:.+$
        connection_ref:
          type: string
          minLength: 1
        entity_ref:
          type: string
          minLength: 1
    TrackerLinkResponse:
      type: object
      additionalProperties: false
      required: [typed_ref, external_refs, sync_event_ref]
      properties:
        typed_ref:
          type: string
          pattern: ^[a-z0-9_-]+:[a-z0-9_-]+:[a-z0-9_-]+:.+$
        external_refs:
          type: array
          items:
            $ref: '#/components/schemas/ExternalRef'
        sync_event_ref:
          type: string
          minLength: 1
    ResultApplyResponse:
      type: object
      additionalProperties: false
      required: [task, emitted_events, next_action]
      properties:
        task:
          $ref: '#/components/schemas/Task'
        emitted_events:
          type: array
          items:
            $ref: '#/components/schemas/StateTransitionEvent'
        next_action:
          type: string
          enum: [dispatch_dev, dispatch_acceptance, integrate, publish, wait_manual, none]
    IntegrateRequest:
      type: object
      additionalProperties: false
      required: [expected_state, base_sha]
      properties:
        expected_state:
          type: string
          enum: [accepted]
        base_sha:
          type: string
          pattern: ^[a-fA-F0-9]{7,64}$
        branch_ref:
          type: object
          additionalProperties: true
        patch_ref:
          type: object
          additionalProperties: true
        expected_version:
          type: integer
          minimum: 0
    IntegrateResponse:
      type: object
      additionalProperties: false
      required: [task_id, state, integration_branch]
      properties:
        task_id:
          type: string
        state:
          type: string
          enum: [integrating]
        integration_branch:
          type: string
        run_id:
          type: string
    PublishRequest:
      type: object
      additionalProperties: false
      required: [mode, idempotency_key]
      properties:
        mode:
          type: string
          enum: [no_op, dry_run, apply]
        idempotency_key:
          type: string
          minLength: 1
        approval_token:
          type: string
        expected_version:
          type: integer
          minimum: 0
    PublishResponse:
      type: object
      additionalProperties: false
      required: [task_id, state, publish_run_id]
      properties:
        task_id:
          type: string
        state:
          type: string
          enum: [publishing, publish_pending_approval]
        publish_run_id:
          type: string
    JobStatusResponse:
      type: object
      additionalProperties: false
      required: [job_id]
      properties:
        job_id:
          type: string
        task_id:
          type: string
        job:
          $ref: '#/components/schemas/WorkerJob'
        latest_result:
          $ref: '#/components/schemas/WorkerResult'
        lease_expires_at:
          type: string
          format: date-time
        last_heartbeat_at:
          type: string
          format: date-time
    HeartbeatRequest:
      $ref: ./schemas/heartbeat-request.schema.json
    HeartbeatResponse:
      $ref: ./schemas/heartbeat-response.schema.json
    CompleteAcceptanceRequest:
      type: object
      additionalProperties: false
      properties:
        checked_items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id]
            properties:
              id:
                type: string
                minLength: 1
              checked_by:
                type: string
              notes:
                type: string
        verdict:
          type: object
          additionalProperties: false
          required: [outcome]
          properties:
            outcome:
              type: string
              enum: [accept, reject, rework, needs_manual_review]
            reason:
              type: string
            manual_notes:
              type: string
    CompleteAcceptanceResponse:
      type: object
      additionalProperties: false
      required: [task_id, state, checklist_complete]
      properties:
        task_id:
          type: string
        state:
          type: string
          enum: [accepted]
        checklist_complete:
          type: boolean
        verdict_outcome:
          type: string
          enum: [accept, reject, rework, needs_manual_review]
    AuditEvent:
      type: object
      additionalProperties: false
      required: [event_id, event_type, task_id, actor_type, actor_id, payload, occurred_at]
      properties:
        event_id:
          type: string
        event_type:
          type: string
          enum:
            - state_transition
            - job_started
            - job_completed
            - retry_triggered
            - heartbeat_missed
            - orphan_detected
            - lock_conflict
            - capability_mismatch
            - doom_loop_detected
            - approval_requested
            - approval_granted
            - approval_denied
            - run.main_updated
            - run.publishRequested
            - run.publishCompleted
            - task.verdictSubmitted
            - run.permissionEscalated
        task_id:
          type: string
        run_id:
          type: string
        job_id:
          type: string
        actor_type:
          type: string
          enum: [control_plane, worker, human, policy_engine, system]
        actor_id:
          type: string
        payload:
          type: object
          additionalProperties: true
        occurred_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true